pyproject.toml
[tool.poetry]
name = "telegram-standup-bot"
version = "0.1.0"
description = "Telegram bot to do standups"
authors = ["zhukovgreen <zhukovgreen@ya.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^1.4"
envparse = "^0.2.0"
attrs = "^18.2"
valid8 = "^3.7"
mini_lambda = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.9"
neovim = "^0.2.6"
notebook = "^5.7"

[tool.black]
line-length = 54
py37 = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

